requirements.txt
aiofiles==24.1.0
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.6.2.post1
arabic-reshaper==3.0.0
attrs==24.2.0
beautifulsoup4==4.12.3
blend_modes==2.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
distro==1.9.0
et_xmlfile==2.0.0
etelemetry==0.3.1
filelock==3.16.1
fire==0.7.0
fonttools==4.55.0
frontend==0.0.3
fsspec==2024.12.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-vision==3.9.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.36.0
imgaug==0.4.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.0
numpy==1.26.4
openai==1.58.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
paddleocr==2.9.1
pandas==2.2.3
pathlib==1.0.1
pickle-mixin==1.0.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.2
prov==2.0.1
psutil==6.1.1
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydot==3.0.2
pygame==2.6.1
PyMuPDF==1.24.13
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PySocks==1.7.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytils==0.4.1
pytweening==1.2.0
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
RapidFuzz==3.10.1
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.26.0
setuptools==75.6.0
shapely==2.0.6
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.2
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.2.1
tools==0.1.9
torch==2.5.1
tqdm==4.67.1
traits==6.4.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
